+++ src/autograd.js
@@ -18,6 +18,7 @@ import { transpose as gpuTranspose, inverseAxes } from './ops/transpose.js'
 import { reshape as gpuReshape } from './ops/reshape.js'
 import { flashAttentionForward as gpuFlashFwd, flashAttentionBackward as gpuFlashBwd } from './ops/flash_attention.js'
 import { conv2dForward as gpuConv2dFwd, conv2dBackwardInput as gpuConv2dBwdInput, conv2dBackwardWeight as gpuConv2dBwdWeight, conv2dBackwardBias as gpuConv2dBwdBias, convOutputSize } from './ops/conv2d.js'
+import { canUseWinograd, winogradForward as gpuWinogradFwd, winogradBackwardInput as gpuWinogradBwdInput } from './ops/conv2d_winograd.js'
 import { maxPool2dForward as gpuMaxPool2dFwd, maxPool2dBackward as gpuMaxPool2dBwd, avgPool2dForward as gpuAvgPool2dFwd, avgPool2dBackward as gpuAvgPool2dBwd } from './ops/pool2d.js'
 import { batchnormForward as gpuBnFwd, batchnormInference as gpuBnInfer, batchnormBackward as gpuBnBwd, createBatchNorm } from './ops/batchnorm.js'
 import { ropeForward as gpuRopeFwd, ropeBackward as gpuRopeBwd, precomputeRoPE } from './ops/rope.js'
@@ -368,11 +369,15 @@ function crossEntropy(logits, targets) {
 // --- Conv2d ---
 
 function conv2d(input, weight, bias, opts = {}) {
-  const outData = gpuConv2dFwd(input.data, weight.data, bias ? bias.data : null, opts)
+  const useWinograd = canUseWinograd(weight.data, opts)
+  const fwd = useWinograd ? gpuWinogradFwd : gpuConv2dFwd
+  const bwdInput = useWinograd ? gpuWinogradBwdInput : gpuConv2dBwdInput
+  const outData = fwd(input.data, weight.data, bias ? bias.data : null, opts)
   return variable(outData, {
     _deps: bias ? [input, weight, bias] : [input, weight],
     _backward: _noGrad ? null : (grad) => {
-      addGrad(input, gpuConv2dBwdInput(grad, weight.data, input.data.shape, opts))
+      addGrad(input, bwdInput(grad, weight.data, input.data.shape, opts))
+      // Weight gradient always uses direct conv (Winograd weight grad is complex, not worth it)
       addGrad(weight, gpuConv2dBwdWeight(input.data, grad, weight.data.shape, opts))
       if (bias) addGrad(bias, gpuConv2dBwdBias(grad))
     },